Server in Server.scala and ServiceMain in ServiceMain.scala
